+++ assets/js/models/assignments/useAssignmentsCount.tsx
@@ -1,0 +1,11 @@
+import { useGetAssignmentsCount } from "@/api";
+import { useAssignmentsCount as useAssignmentCountChangeSignal } from "@/api/socket";
+
+
+export function useAssignmentsCount() {
+  const { data, refetch } = useGetAssignmentsCount({});
+  
+  useAssignmentCountChangeSignal(refetch);
+
+  return data?.count || 0;
+}
